requirements.txt
Django==4.0.5
django-crispy-forms==1.14.0
djangorestframework==3.13.1
